isola funções utils que verifica, permissões de acesso e de ação
